Move webcam-capture to own directory
